pyproject.toml
[tool.poetry]
name = "Leader-team-AI-bot"
version = "0.1.0"
description = ""
authors = ["ms <ff9147782@gmail.com> <https://github.com/MalakhovStas>"]
repository = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pytoml = "^0.1.21"

redis = "^5.0.1"
aioredis = "2.0.1"
types-redis = "^4.6.0.11"

django = "^4.2.7"
django-cors-headers = "^4.3.1"
django-jinja = "^2.11.0"
django-phonenumber-field = "^7.2.0"
django-redis = "^5.4.0"
djangorestframework = "^3.14.0"

celery = "^5.3.6"
celery-progress = "^0.3"
celery-types = "^0.20.0"
django-celery-results = "^2.5.1"

loguru = "^0.7.2"
phonenumberslite = "^8.13.25"
pillow = "^10.1.0"

aiogram = "2.25.1"
peewee = "3.17.0"
aiohttp = "3.8.6"
aiohttp-proxy = "0.1.2"
openpyxl = "3.1.2"
psutil = "5.9.6"
openai = "0.27.2"
psycopg2-binary = "^2.9.9"

art = "^6.1"
colorama = "^0.4.6"
types-colorama = "^0.4.15.12"
aiofiles = "^23.2.1"



[tool.poetry.group.development.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pylint = "^3.0.2"
django-debug-toolbar = "^4.2.0"
mypy = "^1.6.1"


[tool.poetry.scripts]
somecommand = "-c 'print(1)'"
runserver = "python BlankDjangoAiogramTgBot.manage runsrver"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
pretty = true
#ignore_missing_imports = true

#[[tool.mypy.overrides]] пример
#module = "src.tools.requests_manager"
#disable_error_code = ["assignment", "arg-type", "import-untyped"]